azure_func/requirements.txt
azure-functions
requests==2.32.3
azure-storage-blob==12.19.1
pytz